requirements.txt
# dependencies
colorama
python-dotenv
uvicorn
pydantic>=2,<3
fastapi
aiokafka
SQLAlchemy
PyJWT
psycopg2
redis
structlog
pydantic_settings
langchain_scrapegraph
python-dotenv
uvicorn

# === LANGCHAIN ECOSYSTEM ===
# Versiones compatibles con LangChain 0.3.x
langchain>=0.3,<0.4
langchain-openai>=0.2,<0.3
langsmith
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain_anthropic>=0.2,<0.3
langchain_google_vertexai>=2,<3

# === LANGGRAPH ===
# Versión más reciente estable
langgraph>=0.4.10
langgraph-cli[inmem]>=0.3.3
langgraph-checkpoint-postgres
langgraph-prebuilt 
# === OTRAS DEPENDENCIAS ===
motor
pandas
starlette
httpx
asyncpg
PyPDF2
pydantic[email]
python-multipart
passlib[bcrypt]
pypdf
tavily-python
poetry-core
poetry
psycopg
psycopg_pool
langchain_qdrant
semantic-router
google-api-python-client
typing_extensions